+++ CHANGELOG.md
@@ -7,4 +7,6 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [0.0.1] - 2024-10-24
 
 - Initial release
-- R language support
+- R dataframes support
+- Python dataframes support
+- Julia dataframes support